+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # docs
 
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies [0594176]
+  - @sopt-makers/icons@1.1.0
+  - @sopt-makers/ui@2.8.3
+
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [161f23a]
+  - @sopt-makers/ui@2.8.2
+
 ## 0.0.26
 
 ### Patch Changes